Adding indexer for user all-time action summary

+++ packages/common/src/main/kotlin/org/vechain/indexer/b3tr/shared/UserActionSummaryDocument.kt
@@ -1,11 +1,12 @@
 package org.vechain.indexer.b3tr.shared
 
+import java.math.BigDecimal
 import org.vechain.indexer.VersionedDocument
 import org.vechain.indexer.b3tr.action.Impact
 
 interface UserActionSummaryDocument : VersionedDocument {
     val entity: String
     val actionsRewarded: Long
-    val totalRewardAmount: Double
+    val totalRewardAmount: BigDecimal
     val totalImpact: Impact?
 }
